+++ src/pages/Home/index.tsx
@@ -1,11 +1,19 @@
-import { useEffect } from '@/libs/valueToUI'
+import { useAPI } from '@/hooks/useAPI'
 
 const Home = () => {
-  useEffect(() => {
-    fetch('/api/user/1')
-      .then((res) => res.json())
-      .then(console.log)
-  }, [])
+  const { data, isLoading, isSuccess } = useAPI<{
+    id: string
+    firstName: string
+    age: number
+  }>('get', '/user/1')
+
+  if (isLoading) {
+    return <div>loading...</div>
+  }
+
+  if (!isSuccess) {
+    return <div>error</div>
+  }
 
   const handleClick = async () => {
     const res = await fetch('/api/user/2')
@@ -15,6 +23,7 @@ const Home = () => {
 
   return (
     <div>
+      {JSON.stringify(data)}
       <button onclick={handleClick}>눌러봐</button>
     </div>
   )
